organize, rename, couple bug fix
