+++ code/scripts/graphoutlier.py
@@ -4,14 +4,17 @@ and meanSignal per subject per run and indicates potential outliers (motion-
 induced artifacts) based on commonly used threshold of 0.2-0.5mm for FD and 
 0.3-0.5% for DVARS. 0.5 was used for DVARS and 0.5 was used for FD.
 """
-import numpy as np
-import nibabel as nib
-import matplotlib.pyplot as plt
-import graph_functions as gf
+import sys
 import json
-
+# Paths
 pathtodata = '../../data/'
 pathtofig = '../../paper/figures/'
+pathtofunction = '../utils'
+# Append fuction path
+sys.path.append(pathtofunction)
+# Import function
+from graphoutlier_functions import loadnib_dict, loadtxt_dict, plot_dvars, plot_fd, plot_meanSig
+
 
 # load outlier files
 dvars_out = json.load(open(pathtodata+"ds005/dvarsOutliers.txt"))
@@ -28,21 +31,21 @@ for i in range(1,10):
         # dvars path and name, call function
         dvarsfile=txtpath+'dvars.txt'
         dvarsfigname=pathtofig+'dvars_sub'+`i`+'run'+`j`+'.png'
-        dvars_dict = gf.loadtxt_dict(dvarsfile, dvarsfigname)
+        dvars_dict = loadtxt_dict(dvarsfile, dvarsfigname)
         dvars_outliers = dvars_out['sub'+`i`+'run'+`j`]
-        gf.plot_dvars(dvars_dict, dvars_outliers, saveit=True)
+        plot_dvars(dvars_dict, dvars_outliers, saveit=True)
         # fd path and name, call function
         fdfile=txtpath+'fd.txt'
         fdfigname=pathtofig+'fd_sub'+`i`+'run'+`j`+'.png'
-        fd_dict = gf.loadtxt_dict(fdfile, fdfigname)
+        fd_dict = loadtxt_dict(fdfile, fdfigname)
         fd_outliers = fd_out['sub'+`i`+'run'+`j`]
-        gf.plot_fd(fd_dict, fd_outliers, saveit=True)
+        plot_fd(fd_dict, fd_outliers, saveit=True)
         # mean path and name, call function
         niipath=pathtodata+'ds005/sub00'+`i`+'/BOLD/task001_run00'+`j`
         meandata=niipath+'/bold.nii.gz'
         meanfigname=pathtofig+'mean_sub'+`i`+'run'+`j`+'.png'
-        data_dict = gf.loadnib_dict(meandata, meanfigname)
-        gf.plot_meanSig(data_dict, saveit=True)
+        data_dict = loadnib_dict(meandata, meanfigname)
+        plot_meanSig(data_dict, saveit=True)
 
 for i in range(10,17):
     for j in range(1,4):
@@ -52,19 +55,19 @@ for i in range(10,17):
         # dvars path and name, call function
         dvarsfile=txtpath+'dvars.txt'
         dvarsfigname=pathtofig+'dvars_sub'+`i`+'run'+`j`+'.png'
-        dvars_dict = gf.loadtxt_dict(dvarsfile, dvarsfigname)
+        dvars_dict = loadtxt_dict(dvarsfile, dvarsfigname)
         dvars_outliers = dvars_out['sub'+`i`+'run'+`j`]
-        gf.plot_dvars(dvars_dict, dvars_outliers, saveit=True)
+        plot_dvars(dvars_dict, dvars_outliers, saveit=True)
         # fd path and name, call function
         fdfile=txtpath+'fd.txt'
         fdfigname=pathtofig+'fd_sub'+`i`+'run'+`j`+'.png'
-        fd_dict = gf.loadtxt_dict(fdfile, fdfigname)
+        fd_dict = loadtxt_dict(fdfile, fdfigname)
         fd_outliers = fd_out['sub'+`i`+'run'+`j`]
-        gf.plot_fd(fd_dict, fd_outliers, saveit=True)
+        plot_fd(fd_dict, fd_outliers, saveit=True)
         # mean path and name, call function
         niipath=pathtodata+'ds005/sub0'+`i`+'/BOLD/task001_run00'+`j`
         meandata=niipath+'/bold.nii.gz'
         meanfigname=pathtofig+'mean_sub'+`i`+'run'+`j`+'.png'
-        data_dict = gf.loadnib_dict(meandata, meanfigname)
-        gf.plot_meanSig(data_dict, saveit=True)
+        data_dict = loadnib_dict(meandata, meanfigname)
+        plot_meanSig(data_dict, saveit=True)
